requirements.txt
Flask>=3.0.2
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Werkzeug>=3.0.1
pyotp>=2.9.0
docxtpl>=0.16.7
qrcode>=7.4.2
python-dotenv>=1.0.1
Pillow>=10.2.0
psutil>=5.9.0
Flask-Executor>=1.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
